requirements.txt
pytest==8.3.5
python-dotenv==1.0.1
fastapi
uvicorn
mysql-connector-python
